[DR-90] refactor: replace edit route with parameterized edit route
- Remove static edit route
- Add parameterized edit/$artistId route with data fetching
- Update EditButton to use parameterized route
